group specs on server
